chore(Theme): rename keybind_utils to utils
Because `isClickGuiScreen` is not a keybind thing
