+++ src-theme/src/routes/hud/elements/KeyBinds.svelte
@@ -4,7 +4,7 @@
     import {listen} from "../../../integration/ws";
     import {convertToSpacedString, spaceSeperatedNames} from "../../../theme/theme_config";
     import type {Module} from "../../../integration/types";
-    import {UNKNOWN_KEY} from "../../../util/keybind_utils";
+    import {UNKNOWN_KEY} from "../../../util/utils";
     import BindDisplay from "../../clickgui/setting/bind/BindDisplay.svelte";
 
     let modules: Module[] = $state([]);
